FSBUILD-221 - fix several trivial warnings
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@16013 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/mod/applications/mod_esf/mod_esf.c
@@ -136,7 +136,7 @@ SWITCH_STANDARD_APP(bcast_function)
 		goto fail;
 	}
 
-	if (switch_mcast_hops(socket, mcast_ttl) != SWITCH_STATUS_SUCCESS) {
+	if (switch_mcast_hops(socket, (uint8_t)mcast_ttl) != SWITCH_STATUS_SUCCESS) {
 		switch_log_printf(SWITCH_CHANNEL_SESSION_LOG(session), SWITCH_LOG_ERROR, "Mutlicast TTL set failed\n");
 		goto fail;
 	}
